Moving away from half_t for hit phis.

+++ cuda/velo/calculate_phi_and_sort/include/CalculatePhiAndSort.cuh
@@ -16,7 +16,7 @@ namespace velo_calculate_phi_and_sort {
     DEVICE_INPUT(dev_velo_cluster_container_t, char) dev_velo_cluster_container;
     DEVICE_OUTPUT(dev_sorted_velo_cluster_container_t, char) dev_sorted_velo_cluster_container;
     DEVICE_OUTPUT(dev_hit_permutation_t, uint) dev_hit_permutation;
-    DEVICE_OUTPUT(dev_hit_phi_t, half_t) dev_hit_phi;
+    DEVICE_OUTPUT(dev_hit_phi_t, float) dev_hit_phi;
     PROPERTY(block_dim_t, DeviceDimensions, "block_dim", "block dimensions", {64, 1, 1});
   };
 
@@ -24,9 +24,9 @@ namespace velo_calculate_phi_and_sort {
     const uint* module_hitStarts,
     const uint* module_hitNums,
     Velo::ConstClusters& velo_cluster_container,
-    half_t* hit_Phis,
+    float* hit_Phis,
     uint* hit_permutations,
-    half_t* shared_hit_phis);
+    float* shared_hit_phis);
 
   __device__ void sort_by_phi(
     const uint event_hit_start,
